refactor by moving smartbulb and smartplug to their own classes
