+++ pyHS100/__init__.py
@@ -1,4 +1,7 @@
 # flake8: noqa
 from __future__ import absolute_import
 from __future__ import unicode_literals
-from pyHS100.pyHS100 import SmartBulb, SmartPlug, TPLinkSmartHomeProtocol, SmartPlugException
+from .smartplug import SmartPlug
+from .pyHS100 import SmartPlugException, SmartDevice
+from .smartbulb import SmartBulb
+from .protocol import TPLinkSmartHomeProtocol
