+++ pyHS100/pyHS100.py
@@ -17,7 +17,9 @@ http://www.apache.org/licenses/LICENSE-2.0
 # python2 compatibility
 from __future__ import absolute_import
 from __future__ import unicode_literals
+
 from future.utils import raise_from
+
 try:
     basestring
 except NameError:
@@ -27,7 +29,7 @@ import datetime
 import logging
 import socket
 
-from pyHS100.protocol import TPLinkSmartHomeProtocol
+from .protocol import TPLinkSmartHomeProtocol
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -405,401 +407,3 @@ class SmartDevice(object):
             return response['power_mw']
         else:
             return response['power']
-
-
-class SmartBulb(SmartDevice):
-    """Representation of a TP-Link Smart Bulb.
-
-    Usage example when used as library:
-    p = SmartBulb("192.168.1.105")
-    # print the devices alias
-    print(p.alias)
-    # change state of bulb
-    p.state = "ON"
-    p.state = "OFF"
-    # query and print current state of plug
-    print(p.state)
-    # check whether the bulb supports color changes
-    if p.is_color:
-    # set the color to an HSV tuple
-    p.hsv = (100, 0, 255)
-    # get the current HSV value
-    print(p.hsv)
-    # check whether the bulb supports setting color temperature
-    if p.is_variable_color_temp:
-    # set the color temperature in Kelvin
-    p.color_temp = 3000
-    # get the current color temperature
-    print(p.color_temp)
-    # check whether the bulb is dimmable
-    if p.is_dimmable:
-    # set the bulb to 50% brightness
-    p.brightness = 50
-    # check the current brightness
-    print(p.brightness)
-
-    Errors reported by the device are raised as SmartPlugExceptions,
-    and should be handled by the user of the library.
-
-    """
-    # bulb states
-    BULB_STATE_ON = 'ON'
-    BULB_STATE_OFF = 'OFF'
-
-    def __init__(self, ip_address, protocol=None):
-        SmartDevice.__init__(self, ip_address, protocol)
-        self.emeter_type = "smartlife.iot.common.emeter"
-        self.emeter_units = True
-
-    @property
-    def is_color(self):
-        """
-        Whether the bulb supports color changes
-
-        :return: True if the bulb supports color changes, False otherwise
-        :rtype: bool
-        """
-        return bool(self.sys_info['is_color'])
-
-    @property
-    def is_dimmable(self):
-        """
-        Whether the bulb supports brightness changes
-
-        :return: True if the bulb supports brightness changes, False otherwise
-        :rtype: bool
-        """
-        return bool(self.sys_info['is_dimmable'])
-
-    @property
-    def is_variable_color_temp(self):
-        """
-        Whether the bulb supports color temperature changes
-
-        :return: True if the bulb supports color temperature changes, False
-        otherwise
-        :rtype: bool
-        """
-        return bool(self.sys_info['is_variable_color_temp'])
-
-    def get_light_state(self):
-        return self._query_helper("smartlife.iot.smartbulb.lightingservice",
-                                  "get_light_state")
-
-    def set_light_state(self, state):
-        return self._query_helper("smartlife.iot.smartbulb.lightingservice",
-                                  "transition_light_state", state)
-
-    @property
-    def hsv(self):
-        """
-        Returns the current HSV state of the bulb, if supported
-
-        :return: tuple containing current hue, saturation and value (0-255)
-        :rtype: tuple
-        """
-
-        if not self.is_color:
-            return None
-
-        light_state = self.get_light_state()
-        if not self.is_on:
-            hue = light_state['dft_on_state']['hue']
-            saturation = light_state['dft_on_state']['saturation']
-            value = int(light_state['dft_on_state']['brightness'] * 255 / 100)
-        else:
-            hue = light_state['hue']
-            saturation = light_state['saturation']
-            value = int(light_state['brightness'] * 255 / 100)
-
-        return(hue, saturation, value)
-
-    @hsv.setter
-    def hsv(self, state):
-        """
-        Sets new HSV, if supported
-
-        :param tuple state: hue, saturation and value (0-255 each)
-        """
-        if not self.is_color:
-            return None
-
-        light_state = {
-            "hue": state[0],
-            "saturation": state[1],
-            "brightness": int(state[2] * 100 / 255),
-            "color_temp": 0
-            }
-        return self.set_light_state(light_state)
-
-    @property
-    def color_temp(self):
-        """
-        Color temperature of the device, if supported
-
-        :return: Color temperature in Kelvin
-        :rtype: int
-        """
-        if not self.is_variable_color_temp:
-            return None
-
-        light_state = self.get_light_state()
-        if not self.is_on:
-            return(light_state['dft_on_state']['color_temp'])
-        else:
-            return(light_state['color_temp'])
-
-    @color_temp.setter
-    def color_temp(self, temp):
-        """
-        Set the color temperature of the device, if supported
-
-        :param int temp: The new color temperature, in Kelvin
-        """
-        if not self.is_variable_color_temp:
-            return None
-
-        light_state = {
-            "color_temp": temp,
-        }
-        return self.set_light_state(light_state)
-
-    @property
-    def brightness(self):
-        """
-        Current brightness of the device, if supported
-
-        :return: brightness in percent
-        :rtype: int
-        """
-        if not self.is_dimmable:
-            return None
-
-        light_state = self.get_light_state()
-        if not self.is_on:
-            return(light_state['dft_on_state']['brightness'])
-        else:
-            return(light_state['brightness'])
-
-    @brightness.setter
-    def brightness(self, brightness):
-        """
-        Set the current brightness of the device, if supported
-
-        :param int brightness: brightness in percent
-        """
-        if not self.is_dimmable:
-            return None
-
-        light_state = {
-            "brightness": brightness,
-        }
-        return self.set_light_state(light_state)
-
-    @property
-    def state(self):
-        """
-        Retrieve the bulb state
-
-        :returns: one of
-                  BULB_STATE_ON
-                  BULB_STATE_OFF
-        :rtype: str
-        """
-        light_state = self.get_light_state()
-        if light_state['on_off']:
-            return self.BULB_STATE_ON
-        return self.BULB_STATE_OFF
-
-    @property
-    def is_on(self):
-        return self.state == self.BULB_STATE_ON
-
-    @state.setter
-    def state(self, bulb_state):
-        """
-        Set the new bulb state
-
-        :param bulb_state: one of
-                           BULB_STATE_ON
-                           BULB_STATE_OFF
-        """
-        print(bulb_state)
-        print(self.BULB_STATE_ON)
-        print(self.BULB_STATE_OFF)
-        if bulb_state == self.BULB_STATE_ON:
-            bulb_state = 1
-        elif bulb_state == self.BULB_STATE_OFF:
-            bulb_state = 0
-        else:
-            raise ValueError
-        
-        light_state = {
-            "on_off": bulb_state,
-        }
-        return self.set_light_state(light_state)
-
-    @property
-    def has_emeter(self):
-        return True
-
-
-class SmartPlug(SmartDevice):
-    """Representation of a TP-Link Smart Switch.
-
-    Usage example when used as library:
-    p = SmartPlug("192.168.1.105")
-    # print the devices alias
-    print(p.alias)
-    # change state of plug
-    p.state = "ON"
-    p.state = "OFF"
-    # query and print current state of plug
-    print(p.state)
-
-    Errors reported by the device are raised as SmartPlugExceptions,
-    and should be handled by the user of the library.
-
-    Note:
-    The library references the same structure as defined for the D-Link Switch
-    """
-    # switch states
-    SWITCH_STATE_ON = 'ON'
-    SWITCH_STATE_OFF = 'OFF'
-    SWITCH_STATE_UNKNOWN = 'UNKNOWN'
-
-    # possible device features
-    FEATURE_ENERGY_METER = 'ENE'
-    FEATURE_TIMER = 'TIM'
-
-    ALL_FEATURES = (FEATURE_ENERGY_METER, FEATURE_TIMER)
-
-    def __init__(self, ip_address, protocol=None):
-        SmartDevice.__init__(self, ip_address, protocol)
-        self.emeter_type = "emeter"
-        self.emeter_units = False
-
-    @property
-    def state(self):
-        """
-        Retrieve the switch state
-
-        :returns: one of
-                  SWITCH_STATE_ON
-                  SWITCH_STATE_OFF
-                  SWITCH_STATE_UNKNOWN
-        :rtype: str
-        """
-        relay_state = self.sys_info['relay_state']
-
-        if relay_state == 0:
-            return SmartPlug.SWITCH_STATE_OFF
-        elif relay_state == 1:
-            return SmartPlug.SWITCH_STATE_ON
-        else:
-            _LOGGER.warning("Unknown state %s returned.", relay_state)
-            return SmartPlug.SWITCH_STATE_UNKNOWN
-
-    @state.setter
-    def state(self, value):
-        """
-        Set the new switch state
-
-        :param value: one of
-                    SWITCH_STATE_ON
-                    SWITCH_STATE_OFF
-        :raises ValueError: on invalid state
-        :raises SmartPlugException: on error
-
-        """
-        if not isinstance(value, basestring):
-            raise ValueError("State must be str, not of %s.", type(value))
-        elif value.upper() == SmartPlug.SWITCH_STATE_ON:
-            self.turn_on()
-        elif value.upper() == SmartPlug.SWITCH_STATE_OFF:
-            self.turn_off()
-        else:
-            raise ValueError("State %s is not valid.", value)
-
-    @property
-    def is_on(self):
-        """
-        Returns whether device is on.
-
-        :return: True if device is on, False otherwise
-        """
-        return bool(self.sys_info['relay_state'])
-
-    def turn_on(self):
-        """
-        Turn the switch on.
-
-        :raises SmartPlugException: on error
-        """
-        self._query_helper("system", "set_relay_state", {"state": 1})
-
-    def turn_off(self):
-        """
-        Turn the switch off.
-
-        :raises SmartPlugException: on error
-        """
-        self._query_helper("system", "set_relay_state", {"state": 0})
-
-    @property
-    def has_emeter(self):
-        """
-        Checks feature list for energey meter support.
-
-        :return: True if energey meter is available
-                 False if energymeter is missing
-        """
-        return SmartPlug.FEATURE_ENERGY_METER in self.features
-
-    @property
-    def features(self):
-        """
-        Returns features of the devices
-
-        :return: list of features
-        :rtype: list
-        """
-        features = self.sys_info['feature'].split(':')
-
-        for feature in features:
-            if feature not in SmartPlug.ALL_FEATURES:
-                _LOGGER.warning("Unknown feature %s on device %s.",
-                                feature, self.model)
-
-        return features
-
-    @property
-    def led(self):
-        """
-        Returns the state of the led.
-
-        :return: True if led is on, False otherwise
-        :rtype: bool
-        """
-        return bool(1 - self.sys_info["led_off"])
-
-    @led.setter
-    def led(self, state):
-        """
-        Sets the state of the led (night mode)
-
-        :param bool state: True to set led on, False to set led off
-        :raises SmartPlugException: on error
-        """
-        self._query_helper("system", "set_led_off", {"off": int(not state)})
-
-    @property
-    def on_since(self):
-        """
-        Returns pretty-printed on-time
-
-        :return: datetime for on since
-        :rtype: datetime
-        """
-        return datetime.datetime.now() - \
-            datetime.timedelta(seconds=self.sys_info["on_time"])
